feat: new reusable Animated header
