+++ src/components/AnimatedFlatList.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
-import { Animated, StyleSheet, useWindowDimensions, View } from 'react-native';
+import { Animated } from 'react-native';
 import type { AnimatedFlatListViewProps } from '../types';
 import AnimatedNavbar from './AnimatedNavbar';
 import { HEADER_HEIGHT, IMG_HEADER_HEIGHT } from '../constants';
 import { useAnimateScrollView } from '../hooks/useAnimateScrollView';
+import { AnimatedHeader } from './AnimatedHeader';
 
 export const AnimatedFlatList = ({
   headerMaxHeight,
@@ -13,47 +14,30 @@ export const AnimatedFlatList = ({
   HeaderNavbarComponent,
   headerImage,
   imageStyle,
+  HeaderComponent,
   ...props
 }: AnimatedFlatListViewProps) => {
-  const { width } = useWindowDimensions();
   const imageHeight = headerMaxHeight || IMG_HEADER_HEIGHT;
   const headerNavHeight = topBarHeight || HEADER_HEIGHT;
   const [scroll, onScroll, scale, translateYDown, translateYUp] =
     useAnimateScrollView(imageHeight, disableScale);
-
-  const ListHeader = () => (
-    <View
-      style={[
-        styles.imgContainer,
-        {
-          marginTop: -imageHeight * 4,
-          paddingTop: imageHeight * 4,
-        },
-      ]}
-    >
-      <Animated.Image
-        source={headerImage}
-        style={[
-          { height: imageHeight, width: width * 1.2 },
-          {
-            transform: [
-              { scale },
-              { translateY: translateYUp },
-              { translateY: translateYDown },
-            ],
-          },
-          imageStyle,
-        ]}
-      />
-    </View>
-  );
 
   return (
     <>
       <Animated.FlatList
         {...props}
         onScroll={onScroll}
-        ListHeaderComponent={ListHeader}
+        ListHeaderComponent={
+          <AnimatedHeader
+            HeaderComponent={HeaderComponent}
+            headerImage={headerImage}
+            imageStyle={imageStyle}
+            imageHeight={imageHeight}
+            translateYDown={translateYDown}
+            translateYUp={translateYUp}
+            scale={scale}
+          />
+        }
       />
       <AnimatedNavbar
         headerHeight={headerNavHeight}
@@ -65,10 +49,3 @@ export const AnimatedFlatList = ({
     </>
   );
 };
-
-const styles = StyleSheet.create({
-  imgContainer: {
-    alignItems: 'center',
-    overflow: 'hidden',
-  },
-});
