+++ src/components/AnimatedScrollView.tsx
@@ -1,15 +1,10 @@
-import {
-  Animated,
-  ImageBackground,
-  StyleSheet,
-  useWindowDimensions,
-  View,
-} from 'react-native';
+import { Animated } from 'react-native';
 import React from 'react';
 import { HEADER_HEIGHT, IMG_HEADER_HEIGHT } from '../constants';
 import AnimatedNavbar from './AnimatedNavbar';
 import type { AnimatedScrollViewProps } from '../types';
 import { useAnimateScrollView } from '../hooks/useAnimateScrollView';
+import { AnimatedHeader } from './AnimatedHeader';
 
 export const AnimatedScrollView = ({
   TopNavBarComponent,
@@ -23,14 +18,10 @@ export const AnimatedScrollView = ({
   imageStyle,
   ...props
 }: AnimatedScrollViewProps) => {
-  const { width } = useWindowDimensions();
   const imageHeight = headerMaxHeight || IMG_HEADER_HEIGHT;
   const headerNavHeight = topBarHeight || HEADER_HEIGHT;
   const [scroll, onScroll, scale, translateYDown, translateYUp] =
     useAnimateScrollView(imageHeight, disableScale);
-
-  const AnimatedImageBackground =
-    Animated.createAnimatedComponent(ImageBackground);
 
   return (
     <>
@@ -39,68 +30,15 @@ export const AnimatedScrollView = ({
         scrollEventThrottle={16}
         {...props}
       >
-        <View
-          style={[
-            styles.imgContainer,
-            {
-              marginTop: -imageHeight * 4,
-              paddingTop: imageHeight * 4,
-            },
-          ]}
-        >
-          {HeaderComponent ? (
-            <>
-              {headerImage ? (
-                <AnimatedImageBackground
-                  source={headerImage}
-                  style={[
-                    { height: imageHeight, width: width * 1.2 },
-                    {
-                      transform: [
-                        { scale },
-                        { translateY: translateYUp },
-                        { translateY: translateYDown },
-                      ],
-                    },
-                    imageStyle,
-                  ]}
-                >
-                  {HeaderComponent}
-                </AnimatedImageBackground>
-              ) : (
-                <Animated.View
-                  style={[
-                    { height: imageHeight, width: width * 1.2 },
-                    {
-                      transform: [
-                        { scale },
-                        { translateY: translateYUp },
-                        { translateY: translateYDown },
-                      ],
-                    },
-                  ]}
-                >
-                  {HeaderComponent}
-                </Animated.View>
-              )}
-            </>
-          ) : (
-            <Animated.Image
-              source={headerImage}
-              style={[
-                { height: imageHeight, width: width * 1.2 },
-                {
-                  transform: [
-                    { scale },
-                    { translateY: translateYUp },
-                    { translateY: translateYDown },
-                  ],
-                },
-                imageStyle,
-              ]}
-            />
-          )}
-        </View>
+        <AnimatedHeader
+          HeaderComponent={HeaderComponent}
+          headerImage={headerImage}
+          imageStyle={imageStyle}
+          imageHeight={imageHeight}
+          translateYDown={translateYDown}
+          translateYUp={translateYUp}
+          scale={scale}
+        />
         {children}
       </Animated.ScrollView>
       <AnimatedNavbar
@@ -113,10 +51,3 @@ export const AnimatedScrollView = ({
     </>
   );
 };
-
-const styles = StyleSheet.create({
-  imgContainer: {
-    alignItems: 'center',
-    overflow: 'hidden',
-  },
-});
